output_modules: cast torch.norm result back to vec1.dtype for bf16-mixed
Second bf16-mixed crash, different site than noise_normalizer (45b0f60).
Under PL --precision bf16-mixed and cuda autocast, the upstream Linear
(self.vec1_proj) produces a bf16 tensor, so `vec1` is allocated bf16
too. `torch.norm` is on autocast's fp32-promotion list and returns fp32
on CUDA — the in-place `vec1[mask] = ...` then mismatches dtypes.

Cast the norm result to vec1.dtype to keep the assignment dtype-clean.

No regression test: CPU autocast(bf16) doesn't reproduce the crash (only
CUDA autocast promotes torch.norm), and there's no cuda_only test
fixture in the repo yet. The inline comment documents the pattern so
this doesn't get re-introduced silently.

+++ models/ET_models/output_modules.py
@@ -132,7 +132,12 @@ class GatedEquivariantBlock(nn.Module):
                     "These atoms will not interact with any other atom unless you change the cutoff."
                 )
             )
-        vec1[mask] = torch.norm(vec1_buffer[mask], dim=-2) # NOTE: this is equivariant restriction on expresivity
+        # `torch.norm` is on autocast's fp32-promotion list, so under
+        # `--precision bf16-mixed` it returns fp32 even on bf16 input. `vec1`
+        # is bf16 (created with `dtype=vec1_buffer.dtype`, and vec1_buffer
+        # comes from the autocast-bf16 Linear above). Cast back to the
+        # destination dtype so the in-place write doesn't raise.
+        vec1[mask] = torch.norm(vec1_buffer[mask], dim=-2).to(vec1.dtype) # NOTE: this is equivariant restriction on expresivity
 
         vec2 = self.vec2_proj(v)
         x = torch.cat([x, vec1], dim=-1) # [B, 4, d]
